account creation guessConfig: Don't add SSL cert overrides anymore, just let current connection pass -- bug 721369, r=bienvenu When we were probing hostnames, and encountered SSL cert errors, we let the current connection pass and also added a cert override, without user confirmation. The idea was to try the host again later with a lower priority and try other hosts first. Unfortunately, the code added it as permanent override, so that But that means the user wouldn't get any warning anymore at later stages, so we didn't warn about man-in-the-middle attacks anymore, which means we removed all SSL protections. Great. To add to it, the overrides were explicitly marked permanent, not temporary.
Instead of playing around with the cert override store, we now
simply allow the current socket connection to proceed despite
cert errors, in guessConfig. When the user then goes to verifyConfig,
he will get the normal PSM cert error dialog with the ability to
confirm and add a permanent override manually.