Bug 849540, part 2: Implement an OAuth2 SASL module helper, r=rkent.
The code added in msgOAuth2Module.js is glue functionality to use the
configuration preferences to run the OAuth2.jsm logon process. It also makes the
process usable from C++, so that it can be consumed by IMAP and SMTP
implementations.

At present, the value for GMail is hardcoded. This is hopefully a temporary
measure until dynamic client configuration is supported, at which point it can
be removed. Unfortunately, this also requires us to hardcode the identifier for
the client instead of using, say, navigator.appName.

CLOSED TREE push for SeaMonkey localization changes.